Fix writing of partial frames for modes other than `i`
Originally, only the `i` mode (instruction count as the timeline) was supported.
Other modes were added later. However, one place (the end timestamp of
stack frames that were in progress when the recording started) was not updated
to respect the mode setting, and is still using instruction count.

This results in a malformed trace when the instruction counter happens to be bigger
than the chosen timeline counter (instructions, cycles, nanoseconds).
I overlooked this bug because I was only tracing programs with low IPC
so far.

Fix this by using the chosen counter in the end timestamp, as it should be.